Fix many errors in the BVH construction and traversal
In the end could not make the compiler accept a code using references.
It seems that combining mutation and recursive functions is difficult.
Some initial search suggests it is not possible r/n, and points to
future compiler improvements that would accept more programs.
We'll see if things are different when we use indirection instead of
manipulating the original triangle list.